few addition for our page allocater. the globalMemTbl indicates host physical addresses

+++ vcproj/kernel/kernel.c
@@ -75,7 +75,7 @@ u64 _cdecl __pa(hva_t virt)
 	PHYSICAL_ADDRESS paddr;
 	paddr = MmGetPhysicalAddress((PVOID)virt);	
 	retVal = PAU64(&paddr);
-	SAFE_ASSERT(retVal != 0x0);
+	SAFE_ASSERT(retVal != (u64)0x0);
 	return retVal;
 }
